backend/requirements.txt
# Core Django
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3

# Django Channels (WebSocket)
channels[daphne]>=4.0
channels-redis>=4.0

# Database - SQLite for dev, PostgreSQL for prod
# psycopg2-binary needed only for production PostgreSQL deployment

# Celery (Async Tasks)
celery>=5.4
redis>=5.0

# ML & Data
scikit-learn>=1.5
xgboost>=2.1
pandas>=2.2
numpy>=1.26
joblib>=1.4
scipy>=1.13

# NLP
spacy>=3.7

# LLM / Agent
langchain>=0.2
langchain-core>=0.2
langchain-community>=0.2
langgraph>=0.2
google-generativeai>=0.6
langchain-google-genai>=2.0

# Utilities
python-decouple>=3.8
django-cors-headers>=4.4
drf-spectacular>=0.27
structlog>=24.2
Pillow>=10.3

# Reporting
reportlab>=4.2

# Email
django-anymail>=11.0

# PostgreSQL (production only)
psycopg2-binary>=2.9
